Refactored out bucket eventlist to its own component and added it to QueryExplorer
