+++ src/views/Bucket.vue
@@ -27,118 +27,24 @@ div
 
   b-alert(variant="warning" show)
     | This timeline is a work in progress. It only shows the last 100 events. Hover to get details.
+
   aw-timeline(type="simple" :event_type="bucket.type", :events="events")
 
   hr
 
-  b-card.event-container(no-block=true)
-    span(slot="header")
-      h4.card-title Events
-      span.pagination-header
-        | Showing {{ events.length }} out of ? events
-      b-button(v-on:click="expandList", size="sm", style="float: right;")
-        span(v-if="!isListExpanded")
-          | Expand list
-        span(v-else)
-          | Condense list
-
-    ul.event-list(v-bind:class="{ 'expand': isListExpanded }")
-      li(v-for="event in events")
-          span.event
-            span.field(v-bind:title="event.timestamp")
-              icon(name="calendar-o")
-              | {{ event.timestamp | friendlytime }}
-            span.field
-              icon(name="clock-o")
-              | {{ event.duration | friendlyduration }}
-            span(v-for="(val, key) in event.data").field
-              icon(name="tags")
-              // TODO: Add some kind of highlighting to key
-              | {{ key }}: {{ val }}
+  aw-eventlist(:events="events")
 
 </template>
 
 <style scoped lang="scss">
 
-$border-color: #ddd;
-
-.card {
-  margin-bottom: 1em;
-
-  .card-title {
-    display: inline-block;
-    margin-bottom: 0;
-    margin-right: 1em;
-  }
-
-  .card-body {
-    padding: 0;
-  }
-}
-
-.event-list {
-  list-style-type: none;
-  padding: 0;
-  border-radius: 3px;
-  height: 25em;
-  overflow-y: auto;
-  white-space: nowrap;
-  margin-bottom: 0px;
-
-  li {
-    border: 0 solid $border-color;
-    border-width: 0 0 1px 0;
-    border-radius: 4px;
-    padding: 2px;
-
-    &:last-child {
-      border-width: 0;
-    }
-  }
-
-  &.expand {
-    height: 100%;
-  }
-}
-
-.event {
-  display: inline-block;
-  padding: 0.3em;
-  clear: both;
-}
-
-.pagination-header {
-  font-size: 12pt;
-  color: #666;
-  margin-bottom: 10px;
-}
-
-.field {
-  margin: 0 5px 0 0;
-  font-size: 11pt;
-  padding: 3px 5px 3px 5px;
-  background-color: #ddd;
-  border: 1px solid #ccc;
-  border-radius: 2px;
-
-  &:last-child {
-    margin-right: 0;
-  }
-}
-
-/* Flips the outer element once, then all direct children once,
-   leaving the scrollbar in the first flipped yet the content correct */
-.scrollbar-flipped, .scrollbar-flipped > * {
-  transform: rotateX(180deg);
-  -ms-transform: rotateX(180deg); /* IE 9 */
-  -webkit-transform: rotateX(180deg); /* Safari and Chrome */
-}
 </style>
 
 <script>
 import awclient from '../awclient.js';
 
 import Timeline from '../visualizations/Timeline.vue';
+import EventList from '../visualizations/EventList.vue';
 
 import 'vue-awesome/icons/tags'
 import 'vue-awesome/icons/clock-o'
@@ -148,13 +54,13 @@ export default {
   name: "Bucket",
   components: {
     "aw-timeline": Timeline,
+    "aw-eventlist": EventList,
   },
   data: () => {
     return {
       id: String,
       bucket: Object,
       events: [],
-      isListExpanded: false,
       eventcount: "?",
     }
   },
@@ -176,11 +82,6 @@ export default {
         this.eventcount = response.data;
       });
     },
-
-    expandList: function() {
-      this.isListExpanded = !this.isListExpanded;
-      console.log("List should be expanding: ", this.isListExpanded);
-    }
   },
   mounted: function() {
     this.id = this.$route.params.id;
