+++ src/views/QueryExplorer.vue
@@ -30,10 +30,12 @@ div
     select.form-control(v-model="vis_method")
       option(value="") Select a visualization method
       option(value="timeline") Timeline
-      option(value="other") Other
+      option(value="eventlist") Event List
 
   div(v-if="vis_method == 'timeline'")
     aw-timeline(type="simple", :event_type="event_type", :events="events")
+  div(v-if="vis_method == 'eventlist'")
+    aw-eventlist(:events="events")
 </template>
 
 <style scoped lang="scss">
@@ -44,13 +46,17 @@ import moment from 'moment';
 import awclient from '../awclient.js';
 
 import Timeline from '../visualizations/Timeline.vue';
+import EventList from '../visualizations/EventList.vue';
 
 let today = moment().startOf("day");
 let tomorrow = moment(today).add(24, "hours");
 
 export default {
   name: "QueryExplorer",
-  components: {"aw-timeline": Timeline},
+  components: {
+    "aw-timeline": Timeline,
+    "aw-eventlist": EventList,
+  },
   data: () => {
     return {
       "query_code": "bucketname='aw-watcher-window_hostname';\n\
